Add doExpandUserIcon.js
